# docs/swagger.yaml
basePath: /api
definitions:
  apperror.AppError:
    properties:
      code:
        type: integer
      error_code:
        description: Machine-readable error code
        type: string
      message:
        type: string
    type: object
...
  dto.ApprovalDecisionRequest:
    properties:
      reason:
        maxLength: 255
        type: string
      status:
        enum:
        - approved
        - rejected
        type: string
    required:
    - status
    type: object
...
  dto.CreatePermissionRequest:
    properties:
      description:
        maxLength: 255
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - name
    type: object
...
  dto.CreateRoleRequest:
    properties:
      description:
        maxLength: 255
        type: string
      level:
        description: Level 100 reserved for superadmin
        maximum: 99
        minimum: 0
        type: integer
      name:
        maxLength: 50
        minLength: 3
        type: string
      predefined_name:
        description: 'NEW: Android-style name'
        maxLength: 50
        minLength: 3
        type: string
    required:
    - level
    - name
    - predefined_name
    type: object
...
  dto.PermissionResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.RoleResponse:
    properties:
      description:
        type: string
      id:
        type: string
      is_active:
        description: 'NEW: Role activation status'
        type: boolean
      is_system_role:
        description: 'NEW: System role protection'
        type: boolean
      level:
        description: 'NEW: Hierarchy level'
        type: integer
      name:
        type: string
      permissions:
        items:
          type: string
        type: array
      predefined_name:
        description: 'NEW: Android-style name'
        type: string
    type: object
...
  dto.UpdatePermissionRequest:
    properties:
      description:
        maxLength: 255
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - name
    type: object
  dto.UpdateRolePermissionsRequest:
    properties:
      permission_names:
        description: A list of permission names to assign to the role.
        items:
          type: string
        type: array
    required:
    - permission_names
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the API documentation for the Beresin App backend.
  title: Beresin App API
  version: "1.0"
paths:
...
  /admin/permissions:
    get:
      consumes:
      - application/json
      description: Retrieves all available permissions in the system. Requires 'permissions:read'
        permission.
      produces:
      - application/json
      responses:
        "200":
          description: Permissions retrieved successfully
          schema:
            items:
              $ref: '#/definitions/dto.PermissionResponse'
            type: array
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: List all permissions
      tags:
      - Admin
      - Permissions
    post:
      consumes:
      - application/json
      description: Creates a new permission in the system. Requires 'permissions:create'
        permission.
      parameters:
      - description: New Permission Details
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePermissionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Permission created successfully
          schema:
            $ref: '#/definitions/dto.PermissionResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/apperror.AppError'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/apperror.AppError'
        "409":
          description: Permission with that name already exists
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: Create a new permission
      tags:
      - Admin
      - Permissions
  /admin/permissions/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a permission from the system. Cannot delete if permission
        is assigned to any roles. Requires 'permissions:delete' permission.
      parameters:
      - description: Permission ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Permission deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid permission ID
          schema:
            $ref: '#/definitions/apperror.AppError'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/apperror.AppError'
        "404":
          description: Permission not found
          schema:
            $ref: '#/definitions/apperror.AppError'
        "409":
          description: Cannot delete permission assigned to roles
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: Delete a permission
      tags:
      - Admin
      - Permissions
    put:
      consumes:
      - application/json
      description: Updates an existing permission. Requires 'permissions:update' permission.
      parameters:
      - description: Permission ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated Permission Details
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Permission updated successfully
          schema:
            $ref: '#/definitions/dto.PermissionResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/apperror.AppError'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/apperror.AppError'
        "404":
          description: Permission not found
          schema:
            $ref: '#/definitions/apperror.AppError'
        "409":
          description: Permission name already exists
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: Update a permission
      tags:
      - Admin
      - Permissions
  /admin/roles:
    get:
      consumes:
      - application/json
      description: Retrieves all roles with their permissions. Requires 'roles:read'
        permission.
      produces:
      - application/json
      responses:
        "200":
          description: Roles retrieved successfully
          schema:
            items:
              $ref: '#/definitions/dto.RoleResponse'
            type: array
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: List all roles
      tags:
      - Admin
      - Roles
    post:
      consumes:
      - application/json
      description: Creates a new role. Requires 'roles:create' permission.
      parameters:
      - description: New Role Details
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/dto.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Role created successfully
          schema:
            $ref: '#/definitions/dto.RoleResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/apperror.AppError'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/apperror.AppError'
        "409":
          description: Role with that name already exists
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: Create a new role
      tags:
      - Admin
      - Roles
  /admin/roles/{id}/permissions:
    put:
      consumes:
      - application/json
      description: Updates the list of permissions associated with a specific role.
        This action requires 'roles:assign' permission.
      parameters:
      - description: Role ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: List of permission names
        in: body
        name: permissions
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateRolePermissionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Role permissions updated successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid role ID format or validation failed
          schema:
            $ref: '#/definitions/apperror.AppError'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/apperror.AppError'
        "404":
          description: Role not found
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: Update permissions for a role
      tags:
      - Admin
      - Roles
  /admin/roles/approval-requests:
    get:
      consumes:
      - application/json
      description: Retrieves all role approval requests. Requires 'roles:approve'
        permission.
      produces:
      - application/json
      responses:
        "200":
          description: Role approval requests retrieved successfully
          schema:
            items:
              $ref: '#/definitions/dto.RoleApprovalResponse'
            type: array
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: List all role approval requests
      tags:
      - Admin
      - Roles
      - Approval
  /admin/roles/approval-requests/{id}/decision:
    put:
      consumes:
      - application/json
      description: Approves or rejects a role approval request. Requires 'roles:approve'
        permission.
      parameters:
      - description: Approval Request ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Approval Decision
        in: body
        name: decision
        required: true
        schema:
          $ref: '#/definitions/dto.ApprovalDecisionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Role request processed successfully
          schema:
            $ref: '#/definitions/dto.RoleApprovalResponse'
        "400":
          description: Invalid request ID format or validation failed
          schema:
            $ref: '#/definitions/apperror.AppError'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/apperror.AppError'
        "404":
          description: Approval request not found
          schema:
            $ref: '#/definitions/apperror.AppError'
        "409":
          description: Request already processed
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: Approve or reject a role request
      tags:
      - Admin
      - Roles
      - Approval
...
  /roles/approval-requests:
    post:
      consumes:
      - application/json
      description: Creates a new role approval request. Users can request new roles
        to be created.
      parameters:
      - description: Role Approval Request Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateRoleApprovalRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Role approval request created successfully
          schema:
            $ref: '#/definitions/dto.RoleApprovalResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/apperror.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: Create a role approval request
      tags:
      - Roles
      - Approval
  /roles/predefined-options:
    get:
      consumes:
      - application/json
      description: Retrieves available predefined role options with their levels and
        descriptions.
      produces:
      - application/json
      responses:
        "200":
          description: Predefined role options retrieved successfully
          schema:
            items:
              $ref: '#/definitions/dto.PredefinedRoleOption'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: Get predefined role options
      tags:
      - Roles
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"